refactored settlement test helpers; added optional --disable-code-size-limit flag to run-integration-tests.sh
