+++ integration/v2/src/tests/settlement/external_match/mod.rs
@@ -1,13 +1,17 @@
 //! External settlement tests
 
+use crate::{
+    test_args::TestArgs,
+    tests::settlement::{BOUNDED_MAX_AMT, Intent, get_total_fee, random_price},
+    util::transactions::wait_for_tx_success,
+};
 use alloy::primitives::U256;
 use eyre::Result;
+use rand::{Rng, thread_rng};
+use renegade_circuits::test_helpers::create_bounded_match_result_with_balance;
 use renegade_crypto::fields::scalar_to_u128;
+use renegade_darkpool_types::bounded_match_result::BoundedMatchResult;
 use renegade_darkpool_types::{fee::FeeTake, settlement_obligation::SettlementObligation};
-
-use crate::{
-    test_args::TestArgs, tests::settlement::get_total_fee, util::transactions::wait_for_tx_success,
-};
 
 pub mod private_intent_private_balance;
 pub mod private_intent_public_balance;
@@ -62,4 +66,34 @@ pub async fn setup_external_match(args: &TestArgs) -> Result<()> {
     wait_for_tx_success(quote.approve(darkpool, amt)).await?;
 
     Ok(())
+}
+
+/// Create an intent and bounded match result constrained by a balance amount
+///
+/// This ensures `max_internal_party_amount_in <= balance_amount`, satisfying
+/// the circuit's capitalization constraint.
+///
+/// Party 0 (internal party) sells the base; Party 1 (external party) sells the quote
+///
+/// Returns the intent, bounded match result, and balance amount used
+pub(crate) fn create_intent_and_bounded_match_result(
+    args: &TestArgs,
+) -> Result<(Intent, BoundedMatchResult, u128)> {
+    let mut rng = thread_rng();
+    // Generate balance first, then intent amount >= balance for meaningful bounded results
+    let balance_amount = rng.gen_range(1..BOUNDED_MAX_AMT);
+    let amount_in = rng.gen_range(balance_amount..=BOUNDED_MAX_AMT);
+    let min_price = random_price();
+    let internal_party_intent = Intent {
+        in_token: args.base_addr()?,
+        out_token: args.quote_addr()?,
+        owner: args.party0_addr(),
+        min_price,
+        amount_in,
+    };
+
+    // Use create_bounded_match_result_with_balance to ensure max <= balance_amount
+    let bounded_match_result =
+        create_bounded_match_result_with_balance(&internal_party_intent, balance_amount);
+    Ok((internal_party_intent, bounded_match_result, balance_amount))
 }
