+++ integration/v2/src/tests/settlement/private_fill.rs
@@ -37,8 +37,8 @@ use test_helpers::{assert_eq_result, integration_test_async};
 use crate::{
     test_args::TestArgs,
     tests::settlement::{
-        compute_fee_take, create_random_intents_and_obligations,
-        private_intent_private_balance::{self, fund_ring2_party},
+        compute_fee_take, create_random_intents_and_obligations, fund_ring2_party,
+        private_intent_private_balance::{self},
         settlement_relayer_fee, split_obligation,
     },
     util::{merkle::find_state_element_opening, transactions::wait_for_tx_success},
